+++ CHANGELOG.md
@@ -8,9 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line.
 
-
 ### Fixed
 
+- Core: include arbitrary refunds for max refundable amount
 - Admin: select product variation in popup window 
 - Importer: ignore None columns while importing files